requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-filter==23.2
django-ninja==0.20.0
djangorestframework==3.14.0
drf-yasg==1.21.5
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pydantic==1.10.5
pytz==2022.7.1
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.20.0
websockets==11.0.3
